+++ fugitive-darkness-common/src/main/java/com/zer0s2m/fugitivedarkness/common/dto/ContainerGitRepoSearchFilters.java
@@ -14,7 +14,9 @@ public record ContainerGitRepoSearchFilters(
 
         String patternForExcludeFile,
 
-        int maxCount
+        int maxCount,
+
+        int maxDepth
 
 ) {
 }